bpf: rename ENDPOINTS_MAP to cilium_lxc, IPCACHE_MAP to cilium_ipcache
Also moved the map definitions to more-specific headers to allow for more targeted
includes without SKIP_FOO_MAP voodoo.
