+++ bpf/lib/eps.h
@@ -3,10 +3,19 @@
 
 #pragma once
 
-#include <linux/ip.h>
-#include <linux/ipv6.h>
+#include "common.h"
 
-#include "maps.h"
+#include <linux/ip.h>
+#include "ipv6.h"
+
+struct {
+	__uint(type, BPF_MAP_TYPE_HASH);
+	__type(key, struct endpoint_key);
+	__type(value, struct endpoint_info);
+	__uint(pinning, LIBBPF_PIN_BY_NAME);
+	__uint(max_entries, ENDPOINTS_MAP_SIZE);
+	__uint(map_flags, CONDITIONAL_PREALLOC);
+} cilium_lxc __section_maps_btf;
 
 static __always_inline __maybe_unused struct endpoint_info *
 __lookup_ip6_endpoint(const union v6addr *ip6)
@@ -16,7 +25,7 @@ __lookup_ip6_endpoint(const union v6addr *ip6)
 	key.ip6 = *ip6;
 	key.family = ENDPOINT_KEY_IPV6;
 
-	return map_lookup_elem(&ENDPOINTS_MAP, &key);
+	return map_lookup_elem(&cilium_lxc, &key);
 }
 
 static __always_inline __maybe_unused struct endpoint_info *
@@ -33,7 +42,7 @@ __lookup_ip4_endpoint(__u32 ip)
 	key.ip4 = ip;
 	key.family = ENDPOINT_KEY_IPV4;
 
-	return map_lookup_elem(&ENDPOINTS_MAP, &key);
+	return map_lookup_elem(&cilium_lxc, &key);
 }
 
 static __always_inline __maybe_unused struct endpoint_info *
@@ -41,6 +50,32 @@ lookup_ip4_endpoint(const struct iphdr *ip4)
 {
 	return __lookup_ip4_endpoint(ip4->daddr);
 }
+
+struct ipcache_key {
+	struct bpf_lpm_trie_key lpm_key;
+	__u16 cluster_id;
+	__u8 pad1;
+	__u8 family;
+	union {
+		struct {
+			__u32		ip4;
+			__u32		pad4;
+			__u32		pad5;
+			__u32		pad6;
+		};
+		union v6addr	ip6;
+	};
+} __packed;
+
+/* Global IP -> Identity map for applying egress label-based policy */
+struct {
+	__uint(type, BPF_MAP_TYPE_LPM_TRIE);
+	__type(key, struct ipcache_key);
+	__type(value, struct remote_endpoint_info);
+	__uint(pinning, LIBBPF_PIN_BY_NAME);
+	__uint(max_entries, IPCACHE_MAP_SIZE);
+	__uint(map_flags, BPF_F_NO_PREALLOC);
+} cilium_ipcache __section_maps_btf;
 
 /* IPCACHE_STATIC_PREFIX gets sizeof non-IP, non-prefix part of ipcache_key */
 #define IPCACHE_STATIC_PREFIX							\
@@ -92,6 +127,6 @@ ipcache_lookup4(const void *map, __be32 addr, __u32 prefix, __u32 cluster_id)
 }
 
 #define lookup_ip6_remote_endpoint(addr, cluster_id) \
-	ipcache_lookup6(&IPCACHE_MAP, addr, V6_CACHE_KEY_LEN, cluster_id)
+	ipcache_lookup6(&cilium_ipcache, addr, V6_CACHE_KEY_LEN, cluster_id)
 #define lookup_ip4_remote_endpoint(addr, cluster_id) \
-	ipcache_lookup4(&IPCACHE_MAP, addr, V4_CACHE_KEY_LEN, cluster_id)
+	ipcache_lookup4(&cilium_ipcache, addr, V4_CACHE_KEY_LEN, cluster_id)
