+++ bpf/tests/lib/ipcache.h
@@ -18,7 +18,7 @@ __ipcache_v4_add_entry(__be32 addr, __u8 cluster_id, __u32 sec_identity,
 	value.key = spi;
 	value.flag_skip_tunnel = flag_skip_tunnel;
 
-	map_update_elem(&IPCACHE_MAP, &key, &value, BPF_ANY);
+	map_update_elem(&cilium_ipcache, &key, &value, BPF_ANY);
 }
 
 static __always_inline void
@@ -68,7 +68,7 @@ __ipcache_v6_add_entry(const union v6addr *addr, __u8 cluster_id, __u32 sec_iden
 
 	memcpy(&key.ip6, addr, sizeof(*addr));
 
-	map_update_elem(&IPCACHE_MAP, &key, &value, BPF_ANY);
+	map_update_elem(&cilium_ipcache, &key, &value, BPF_ANY);
 }
 
 static __always_inline void
